fix editing actions, add mixin for deletion of array item, fix imports
