+++ library/components/editors/AdditionMixin.js
