+++ library/components/editors/DefaultSubmitter.vue
@@ -1,5 +1,6 @@
 <template lang="pug">
-component(:is="inputComponent" v-model="editedValue" @ok="save()" @cancel="cancel()" @remove="remove()")
+div
+  component(:is="inputComponent" v-model="editedValue" @ok="save()" @cancel="cancel()" @remove="remove()")
   // q-btn(icon="remove" color="primary" size="x-small" dense flat v-if="isArrayItem" @click="remove")
 </template>
